feat: 멘토링 금지 기간 관리 및 커피챗 취소 시 환불 처리 추가
- 멘토링 금지 기간을 지정하기 위한 필드 mentoringBanEndDate 추가
- 커피챗 생성 시 mentoringBanEndDate에 기본값 CURRENT_TIMESTAMP 저장
  - null 값이 들어갈 경우 쿼리 성능 저하(풀스캔) 방지를 위해 기본값 적용
- 멘토가 승인된 커피챗을 1일 전 취소하면 30일간 멘토링 금지 패널티 부여
- 멘토가 커피챗을 취소할 경우 멘티에게 포인트 환불 처리
- 커피챗 조회 시 mentoringBanEndDate < NOW() 조건 추가
  - 멘토링 금지 기간인 커피챗은 조회 제외
- 커피챗 삭제 시, 멘토링 금지 기간이면 COFFEE_CHAT_DELETE_BANNED 예외 발생

+++ boottalk/src/main/java/com/icandoit/boottalk/coffeeChat/service/CoffeeChatInfoService.java
@@ -61,6 +61,12 @@ public class CoffeeChatInfoService {
     @Transactional
     public void deleteMyCoffeeChatInfo(Long userId) {
         CoffeeChatInfo coffeeChatInfo = getCoffeeChatInfoByUserId(userId);
+
+        // 멘토링 활동 금지 당한 커피챗 정보는 삭제 불가
+        if (coffeeChatInfo.isMentoringBanned()) {
+            throw new CustomException(ErrorCode.COFFEE_CHAT_DELETE_BANNED);
+        }
+
         coffeeChatInfoRepository.delete(coffeeChatInfo);
     }
